+++ src/components/favourite/card/FavouriteCard.css
@@ -42,14 +42,23 @@ img {
 
 /* media query */
 
+@media (max-width: 1280px) {
+  .card {
+    width: 15rem;
+    height: 13rem;
+  }
+  .card-image {
+    height: 10rem;
+  }
+}
 @media (max-width: 767px) {
   .card {
-    width: 20rem;
-    height: 13rem;
+    width: 15rem;
+    height: 10rem;
   }
 
   .card-image {
-    height: 10rem;
+    height: 7rem;
   }
 }
 
@@ -64,6 +73,21 @@ img {
   }
 }
 
+@media (max-width: 280px) {
+  .card {
+    width: 15rem;
+    height: 10rem;
+  }
+
+  .card-image {
+    height: 7.5rem;
+  }
+  .card p {
+    font-size: 1rem;
+    font-weight: 400;
+  }
+}
+
 .select {
   position: absolute;
   width: 100%;
@@ -71,8 +95,8 @@ img {
   border-radius: 10px;
   top: 0;
   left: 0;
-  background-color: rgba(0, 0, 0, 0.422);
-  background-image: url(../../assets//Vector.png);
+  background-color: rgba(0, 0, 0, 0.29);
+  background-image: url(../../../assets/Vector.png);
   background-repeat: no-repeat;
   background-position: center;
   background-size: auto;
